refactor: Order 클래스 위치 이동 view -> domain/game
